Fix bug with '=' button and display
Earlier when user pressed '=' button, current operator remain the same
like previous math operation. For example if user pressed 2 + 2, and
then wanted to do 4 - 2, the operator was + sign instead of -, and the
result was 6 instead of 2. Now operator is set to null after pressing
'=' button. In previous commit user has to erase display manually after
setting firstNumber and operator, now when firstNumber is set and user
wants to press new number value of firstNumber is comparing to value of
display and if they are the same, display is set to empty string and
immediately after that pressed number is displayed - which now causing
another bug - when firstNumber for example is set to 36, and user press
365, whole number is replaced by 5.